+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### HEAD
 
   * Added setters for Message#mediaType and Message#charset
+  * Location#concat uses the protocol, auth, and host of its argument, if present
 
 ### 1.0.0-rc3